+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added product specification groups for comparison
+
 ### Fixed
 - fix requested ui issues
 - fix slider refresh issue